+++ src/main/java/http/servlet/HttpServletHandler.java
@@ -1,17 +1,17 @@
-package servlet;
+package http.servlet;
 
 import http.request.HttpRequest;
 import http.response.HttpResponse;
-import servlet.controller.Controller;
-import servlet.controller.ControllerFinder;
+import http.servlet.controller.Controller;
+import http.servlet.controller.ControllerFinder;
 
 import java.io.IOException;
 import java.net.URISyntaxException;
 
-public class HttpRequestHandler {
+public class HttpServletHandler {
     private final ControllerFinder controllerMapper;
 
-    public HttpRequestHandler(final ControllerFinder controllerMapper) {
+    public HttpServletHandler(final ControllerFinder controllerMapper) {
         this.controllerMapper = controllerMapper;
     }
