+++ src/main/java/http/servlet/controller/UserCreateController.java
@@ -1,9 +1,9 @@
-package servlet.controller;
+package http.servlet.controller;
 
 import domain.UserService;
 import http.request.HttpRequest;
 import http.response.HttpResponse;
-import servlet.resolver.UserResolver;
+import http.servlet.resolver.UserResolver;
 
 import java.io.IOException;
