+++ src/main/java/http/servlet/controller/exception/Page404NotFoundException.java
@@ -1,4 +1,4 @@
-package servlet.controller.exception;
+package http.servlet.controller.exception;
 
 public class Page404NotFoundException extends RuntimeException {
     private static final String MESSAGE_OF_404_ERROR = "404 Error\n" + "해당 요청에 대한 페이지를 찾을 수 없습니다.";
