+++ src/main/java/webserver/ServletContainer.java
@@ -1,8 +1,9 @@
-package servlet;
+package webserver;
 
 import file.FileContainer;
 import http.request.HttpRequest;
 import http.response.HttpResponse;
+import http.servlet.HttpServletHandler;
 
 import java.io.IOException;
 import java.net.URISyntaxException;
@@ -10,18 +11,18 @@ import java.net.URISyntaxException;
 public class ServletContainer {
 
     private final FileContainer fileContainer;
-    private final HttpRequestHandler httpRequestHandler;
+    private final HttpServletHandler httpServletHandler;
 
-    public ServletContainer(final FileContainer fileContainer, final HttpRequestHandler httpRequestHandler) {
+    public ServletContainer(final FileContainer fileContainer, final HttpServletHandler httpServletHandler) {
         this.fileContainer = fileContainer;
-        this.httpRequestHandler = httpRequestHandler;
+        this.httpServletHandler = httpServletHandler;
     }
 
     public void process(final HttpRequest httpRequest,
                         final HttpResponse httpResponse) throws IOException, URISyntaxException {
 
         if (!fileContainer.process(httpRequest, httpResponse)) {
-            httpRequestHandler.process(httpRequest, httpResponse);
+            httpServletHandler.process(httpRequest, httpResponse);
         }
     }
 }
