+++ src/main/java/webserver/WebServer.java
@@ -1,10 +1,10 @@
 package webserver;
 
+import http.servlet.controller.*;
 import http.session.support.RandomKeyGenerator;
 import http.session.support.SessionManager;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import servlet.controller.*;
 
 import java.net.ServerSocket;
 import java.net.Socket;
